chore: Refactor pingHandler to accept serverIDs as parameter
The `pingHandler` function has been refactored to accept `serverIDs` as a parameter. This allows for more flexibility in specifying the servers to ping. The `pingHandlerFunc` function now takes `serverIDs` as an argument and returns an `http.HandlerFunc` that handles the ping request. This change improves the modularity and reusability of the code.