feat: Update cronjobs to deploy ping and speed cronjobs
The code changes in the .github/workflows/ci.yml file update the list of manifests to include the new cronjob_ping.yaml and cronjob_speed.yaml files. This ensures that the ping and speed cronjobs are deployed along with the other Kubernetes resources.

The .github/workflows/release.yml file is also updated to deploy the ping and speed cronjobs separately using the azure/k8s-deploy action. The images for the cronjobs are set to use the latest version of the netmon-cli container.

The new cronjob_ping.yaml and cronjob_speed.yaml files define the cronjobs for running the ping and speed commands respectively. They specify the schedule, concurrency policy, successful and failed jobs history limits, and the container configuration. The containers use the latest version of the netmon-cli image and set the OTEL_EXPORTER_OTLP_ENDPOINT environment variable to the Grafana OTLP endpoint.

The k8s/cronjobs.yaml file is deleted as it is no longer needed.

chore: Update Dockerfile to use golang:1.22 AS builder

The Dockerfile is updated to use the `AS` keyword instead of `as` for the builder stage. This change aligns with the established convention of using uppercase for the `AS` keyword in Dockerfiles.

feat: Add support for retrieving speed data by server ID

The code changes modify the requests made to the API endpoints `/api/v1/ping` and `/api/v1/speed` by appending a server ID parameter. This allows the application to retrieve speed data specific to a particular server.

feat: Add environment variable for OTLP endpoint in cronjobs

The code changes add an environment variable `OTEL_EXPORTER_OTLP_ENDPOINT` to the cronjobs.yaml file. This variable is set to "grafana-k8s-monitoring-grafana-agent.observability.svc.cluster.local:4317" and is used by the cronjobs to configure the OTLP endpoint for exporting telemetry data.