feat: Add NETMON_SERVER_URL environment variable to cronjobs
The code changes in the k8s/cronjob_ping.yaml and k8s/cronjob_speed.yaml files add the NETMON_SERVER_URL environment variable to the cronjobs. This variable is set to "netmon-service.netmon.svc.cluster.local:80" and is used by the cronjobs to specify the server URL for the netmon service.